requirements.txt
django>=4.2.9<4.3
psycopg2>=2.8.6,<2.9
debugpy==1.5.1
djangorestframework==3.14.0